refactor: move matcher types into a module
