+++ benches/router_bench.rs
@@ -1,7 +1,8 @@
 use criterion::{BenchmarkId, Criterion, Throughput, criterion_group, criterion_main};
 use rand::prelude::*;
 use regex::Regex;
-use router_prefilter::{Matcher, MatcherVisitor, RouterPrefilter};
+use router_prefilter::RouterPrefilter;
+use router_prefilter::matchers::{Matcher, MatcherVisitor};
 use std::fs;
 use std::hint::black_box;
 use std::sync::LazyLock;
