+++ rustfmt.toml
@@ -1,0 +1,1 @@
+style_edition = "2024"
